% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopic_model.R
\name{scopic_model}
\alias{scopic_model}
\title{Function for fitting topic models to scRNA-Seq gene counts}
\usage{
scopic_model(
  counts.mat = counts.mat,
  type = c("genes", "cells"),
  n.topics = 3,
  n.starts = 5,
  seeds = 1:length(n.starts),
  verbose = TRUE
)
}
\arguments{
\item{counts.mat}{Matrix of scRNA-Seq gene counts (columns are cells, rows are genes)}

\item{type}{Options include "genes" (gene set topics) or "cells" (cell cluster topics)}

\item{n.topics}{Number of topics (transcriptional states)}

\item{n.starts}{Number of random starts (will select best model)}

\item{seeds}{Random seeds (number of seeds must equal n.starts)}

\item{verbose}{TRUE to print status updates; FALSE to silence status updates}
}
\value{
Topic model of scRNA-Seq gene counts
}
\description{
This function fits topic models to scRNA-Seq gene counts
}
\examples{
scopic_model(counts.mat = counts.mat, type = c("genes", "cells"), n.topics = 3, n.starts = 5, seeds = length(n.starts), verbose = TRUE)
}
